backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langgraph>=0.4.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "cryptography>=46.0.0",
    "fastapi>=0.135.2",
    "google-api-python-client>=2.169.0",
    "google-auth>=2.40.0",
    "httpx>=0.28.1",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.2.2",
    "pymysql>=1.1.1",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.42.0",
    "psycopg[binary]>=3.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
